requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-extensions==3.2.3
structlog==24.1.0
celery==5.3.4
redis==5.0.1
django-redis==5.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0